integrated app.js and multer middleware into server.js, can deprecate app.js but leaving for now for reference.
